Split auth back-end service file into multiple files
